extract RngSeed struct to it's separate class. [skip ci]
